+++ src/common/input.ts
@@ -120,13 +120,18 @@ export class Input extends EventEmitter {
     public enableTriggers: boolean = true;
 
     public getScope() {
-        //if no stack use direct for some performance
-        if (this._stack.length === 0)
-            return this.client.variables;
-        if (!this.stack.named && !this.loops.length)
-            return this.client.variables;
         let scope: any = {};
         Object.assign(scope, this.client.variables);
+        ['$selectedword', '$selword', '$selectedurl', '$selurl', '$selectedline',
+            '$selline', '$selected', '$character', '$copied'].forEach((a) => {
+                scope[a] = window[a];
+                scope[a.substr(1)] = window[a];
+            });
+        //if no stack use direct for some performance
+        if (this._stack.length === 0)
+            return scope;
+        if (!this.stack.named && !this.loops.length)
+            return scope;
         if (this.stack.named)
             Object.assign(scope, this.stack.named);
         if (this.loops.length) {
@@ -149,6 +154,27 @@ export class Input extends EventEmitter {
             //not a property, i or repeatnum
             if (!Object.prototype.hasOwnProperty.call(scope, name) || name === 'i' || name === 'repeatnum')
                 continue;
+            switch (name) {
+                case '$selectedword':
+                case '$selword':
+                case '$selectedurl':
+                case '$selurl':
+                case '$selectedline':
+                case '$selline':
+                case '$selected':
+                case '$character':
+                case '$copied':
+                case 'selectedword':
+                case 'selword':
+                case 'selectedurl':
+                case 'selurl':
+                case 'selectedline':
+                case 'selline':
+                case 'selected':
+                case 'character':
+                case 'copied':
+                    continue;
+            }
             //if i to z and the loop exist skip it
             if (name.length === 1 && ll && name.charCodeAt(0) >= 105 && name.charCodeAt(0) < 105 + ll)
                 continue;